fix(useReducedMotion): use state
