+++ packages/vkui/src/lib/animation/useReducedMotion.ts
@@ -6,17 +6,10 @@ import { useIsomorphicLayoutEffect } from '../useIsomorphicLayoutEffect';
 
 export const REDUCE_MOTION_MEDIA_QUERY = 'screen and (prefers-reduced-motion: reduce)';
 
-export const useReducedMotion = (): boolean => {
+export const useReducedMotion = (): boolean | undefined => {
   const { window } = useDOM();
-  const initial = React.useMemo(
-    () =>
-      window
-        ? window.matchMedia(REDUCE_MOTION_MEDIA_QUERY).matches
-        : /* istanbul ignore next: на текущий момент, покрытие данного кейса неинтересно  */
-          false,
-    [window],
-  );
-  const reducedMotion = React.useRef(initial);
+
+  const [reducedMotion, setReducedMotion] = React.useState<boolean | undefined>(() => undefined);
 
   useIsomorphicLayoutEffect(() => {
     /* istanbul ignore if: невозможный кейс (в SSR вызова этой функции не будет) */
@@ -24,15 +17,15 @@ export const useReducedMotion = (): boolean => {
       return;
     }
     const match = window.matchMedia(REDUCE_MOTION_MEDIA_QUERY);
-    reducedMotion.current = match.matches;
+    setReducedMotion(match.matches);
     /* istanbul ignore next: на текущий момент, покрытие данного кейса неинтересно  */
     const handleMediaQueryChange = (event: MediaQueryListEvent) => {
       /* istanbul ignore next */
-      reducedMotion.current = event.matches;
+      setReducedMotion(event.matches);
     };
     matchMediaListAddListener(match, handleMediaQueryChange);
     return () => matchMediaListRemoveListener(match, handleMediaQueryChange);
   }, [window]);
 
-  return reducedMotion.current;
+  return reducedMotion;
 };
